convert-image-to-jpg-and-create-thumbnails/requirements.txt
--extra-index-url https://us-central1-python.pkg.dev/mealhow/mealhow-python/simple/
Pillow==10.*
gcloud-aio-storage==9.0.0
asyncio==3.4.3
openai==0.28.1
aiohttp==3.9.*
functions-framework==3.*
google-cloud-ndb==2.2.2
mealhow-sdk==0.1.5
